requirements.txt
openai==1.12.0
flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
flask-cors==4.0.0
mcp-python==0.1.0
aiohttp==3.9.3 